refactor: Station 관련 도메인을 별도의 패키지로 분리
