+++ src/main/java/nextstep/subway/station/applicaion/StationService.java
@@ -1,9 +1,9 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.station.applicaion;
 
-import nextstep.subway.applicaion.dto.StationRequest;
-import nextstep.subway.applicaion.dto.StationResponse;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
+import nextstep.subway.station.applicaion.dto.StationRequest;
+import nextstep.subway.station.applicaion.dto.StationResponse;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
