+++ src/main/java/nextstep/subway/station/applicaion/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.station.applicaion.dto;
 
 public class StationRequest {
     private String name;
