+++ src/main/java/nextstep/subway/station/applicaion/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.station.applicaion.dto;
 
 public class StationResponse {
     private Long id;
